+++ src/vue/VueUtilisateur.java
@@ -178,10 +178,14 @@ public class VueUtilisateur extends JFrame implements ActionListener {
 					int retour = JOptionPane.showConfirmDialog(null, "Voulez-vous supprimer ce Salarié ?", "Suppression", JOptionPane.YES_NO_OPTION); 
 					if (retour == 0) {
 						//suppression dans la base 
-						Main.deleteUtilisateurSalarie(idUtilisateurSalarie);
-						//suppression dans la table d'affichage 
-						unTableauSalarie.deleteLigne(ligne);
-						JOptionPane.showMessageDialog(null, "Suppression réussie");
+						int resultDeleteSalarie = Main.deleteUtilisateurSalarie(idUtilisateurSalarie);
+						if (resultDeleteSalarie == 0) {
+							//s'il n'y a pas d'erreur, on supprime dans le tableau JTable et on affiche succès
+							//suppression dans la table d'affichage 
+							unTableauSalarie.deleteLigne(ligne);
+							JOptionPane.showMessageDialog(null, "Suppression réussie");
+						}
+
 					}
 				}else if (e.getClickCount() == 1) {
 					//modifier
